@@ -11,8 +11,9 @@ export function wrapCode(magicString, uses, moduleName, exportsName) {
   }
   magicString
     .trim()
+    .indent('\t')
     .prepend(`(function (${args.join(', ')}) {\n`)
-    .append(`\n}(${passedArgs.join(', ')}));`);
+    .append(`\n} (${passedArgs.join(', ')}));`);
 }
 
 export function rewriteExportsAndGetExportsBlock(
@@ -30,12 +31,27 @@ export function rewriteExportsAndGetExportsBlock(
   HELPERS_NAME,
   exportMode,
   detectWrappedDefault,
-  defaultIsModuleExports
+  defaultIsModuleExports,
+  usesRequireWrapper,
+  requireName
 ) {
   const exports = [];
   const exportDeclarations = [];
 
-  if (exportMode === 'replace') {
+  if (usesRequireWrapper) {
+    getExportsWhenUsingRequireWrapper(
+      magicString,
+      wrapped,
+      exportMode,
+      exports,
+      moduleExportsAssignments,
+      exportsAssignmentsByName,
+      moduleName,
+      exportsName,
+      requireName,
+      defineCompiledEsmExpressions
+    );
+  } else if (exportMode === 'replace') {
     getExportsForReplacedModuleExports(
       magicString,
       exports,
@@ -76,6 +92,49 @@ export function rewriteExportsAndGetExportsBlock(
   }
 
   return `\n\n${exportDeclarations.join('\n')}`;
+}
+
+function getExportsWhenUsingRequireWrapper(
+  magicString,
+  wrapped,
+  exportMode,
+  exports,
+  moduleExportsAssignments,
+  exportsAssignmentsByName,
+  moduleName,
+  exportsName,
+  requireName,
+  defineCompiledEsmExpressions
+) {
+  if (!wrapped) {
+    if (exportMode === 'replace') {
+      for (const { left } of moduleExportsAssignments) {
+        magicString.overwrite(left.start, left.end, exportsName);
+      }
+    } else {
+      // Collect and rewrite module.exports assignments
+      for (const { left } of moduleExportsAssignments) {
+        magicString.overwrite(left.start, left.end, `${moduleName}.exports`);
+      }
+      // Collect and rewrite named exports
+      for (const [exportName, { nodes }] of exportsAssignmentsByName) {
+        for (const node of nodes) {
+          magicString.overwrite(node.start, node.left.end, `${exportsName}.${exportName}`);
+        }
+      }
+      // Collect and rewrite exports.__esModule assignments
+      for (const expression of defineCompiledEsmExpressions) {
+        const moduleExportsExpression =
+          expression.type === 'CallExpression' ? expression.arguments[0] : expression.left.object;
+        magicString.overwrite(
+          moduleExportsExpression.start,
+          moduleExportsExpression.end,
+          exportsName
+        );
+      }
+    }
+  }
+  exports.push(`${requireName} as __require`);
 }
 
 function getExportsForReplacedModuleExports(
@@ -165,7 +224,7 @@ function getExports(
   }
 
   if (!isRestorableCompiledEsm || defaultIsModuleExports === true) {
-    exportDeclarations.push(`export default ${exportsName};`);
+    exports.push(`${exportsName} as default`);
   } else if (moduleExportsAssignments.length === 0 || defaultIsModuleExports === false) {
     exports.push(`${deconflictedDefaultExportName || exportsName} as default`);
   } else {
