@@ -1,60 +1,31 @@
-import { readFileSync } from 'fs';
-
-import { DYNAMIC_JSON_PREFIX, HELPERS_ID } from './helpers';
-import { getCommonJSMetaPromise } from './is-cjs';
-import { getName, getVirtualPathForDynamicRequirePath, normalizePathSlashes } from './utils';
-
-// e.g. id === "commonjsHelpers?commonjsRegister"
-export function getSpecificHelperProxy(id) {
-  return `export {${id.split('?')[1]} as default} from "${HELPERS_ID}";`;
-}
+import { HELPERS_ID, IS_WRAPPED_COMMONJS } from './helpers';
+import { capitalize, getName } from './utils';
 
 export function getUnknownRequireProxy(id, requireReturnsDefault) {
   if (requireReturnsDefault === true || id.endsWith('.json')) {
-    return `export {default} from ${JSON.stringify(id)};`;
+    return `export { default } from ${JSON.stringify(id)};`;
   }
   const name = getName(id);
   const exported =
     requireReturnsDefault === 'auto'
-      ? `import {getDefaultExportFromNamespaceIfNotNamed} from "${HELPERS_ID}"; export default /*@__PURE__*/getDefaultExportFromNamespaceIfNotNamed(${name});`
+      ? `import { getDefaultExportFromNamespaceIfNotNamed } from "${HELPERS_ID}"; export default /*@__PURE__*/getDefaultExportFromNamespaceIfNotNamed(${name});`
       : requireReturnsDefault === 'preferred'
-      ? `import {getDefaultExportFromNamespaceIfPresent} from "${HELPERS_ID}"; export default /*@__PURE__*/getDefaultExportFromNamespaceIfPresent(${name});`
+      ? `import { getDefaultExportFromNamespaceIfPresent } from "${HELPERS_ID}"; export default /*@__PURE__*/getDefaultExportFromNamespaceIfPresent(${name});`
       : !requireReturnsDefault
-      ? `import {getAugmentedNamespace} from "${HELPERS_ID}"; export default /*@__PURE__*/getAugmentedNamespace(${name});`
+      ? `import { getAugmentedNamespace } from "${HELPERS_ID}"; export default /*@__PURE__*/getAugmentedNamespace(${name});`
       : `export default ${name};`;
   return `import * as ${name} from ${JSON.stringify(id)}; ${exported}`;
 }
 
-export function getDynamicJsonProxy(id, commonDir) {
-  const normalizedPath = normalizePathSlashes(id.slice(DYNAMIC_JSON_PREFIX.length));
-  return `const commonjsRegister = require('${HELPERS_ID}?commonjsRegister');\ncommonjsRegister(${JSON.stringify(
-    getVirtualPathForDynamicRequirePath(normalizedPath, commonDir)
-  )}, function (module, exports) {
-  module.exports = require(${JSON.stringify(normalizedPath)});
-});`;
-}
-
-export function getDynamicRequireProxy(normalizedPath, commonDir) {
-  return `const commonjsRegister = require('${HELPERS_ID}?commonjsRegister');\ncommonjsRegister(${JSON.stringify(
-    getVirtualPathForDynamicRequirePath(normalizedPath, commonDir)
-  )}, function (module, exports) {
-  ${readFileSync(normalizedPath, { encoding: 'utf8' })}
-});`;
-}
-
-export async function getStaticRequireProxy(
-  id,
-  requireReturnsDefault,
-  esModulesWithDefaultExport,
-  esModulesWithNamedExports,
-  commonJsMetaPromises
-) {
+export async function getStaticRequireProxy(id, requireReturnsDefault, loadModule) {
   const name = getName(id);
-  const commonjsMeta = await getCommonJSMetaPromise(commonJsMetaPromises, id);
-  if (commonjsMeta && commonjsMeta.isCommonJS) {
-    return `export { __moduleExports as default } from ${JSON.stringify(id)};`;
-  } else if (commonjsMeta === null) {
+  const {
+    meta: { commonjs: commonjsMeta }
+  } = await loadModule({ id });
+  if (!commonjsMeta) {
     return getUnknownRequireProxy(id, requireReturnsDefault);
+  } else if (commonjsMeta.isCommonJS) {
+    return `export { __moduleExports as default } from ${JSON.stringify(id)};`;
   } else if (!requireReturnsDefault) {
     return `import { getAugmentedNamespace } from "${HELPERS_ID}"; import * as ${name} from ${JSON.stringify(
       id
@@ -62,10 +33,46 @@ export async function getStaticRequireProxy(
   } else if (
     requireReturnsDefault !== true &&
     (requireReturnsDefault === 'namespace' ||
-      !esModulesWithDefaultExport.has(id) ||
-      (requireReturnsDefault === 'auto' && esModulesWithNamedExports.has(id)))
+      !commonjsMeta.hasDefaultExport ||
+      (requireReturnsDefault === 'auto' && commonjsMeta.hasNamedExports))
   ) {
     return `import * as ${name} from ${JSON.stringify(id)}; export default ${name};`;
   }
   return `export { default } from ${JSON.stringify(id)};`;
+}
+
+export function getEntryProxy(id, defaultIsModuleExports, getModuleInfo) {
+  const {
+    meta: { commonjs: commonjsMeta },
+    hasDefaultExport
+  } = getModuleInfo(id);
+  if (!commonjsMeta || commonjsMeta.isCommonJS !== IS_WRAPPED_COMMONJS) {
+    const stringifiedId = JSON.stringify(id);
+    let code = `export * from ${stringifiedId};`;
+    if (hasDefaultExport) {
+      code += `export { default } from ${stringifiedId};`;
+    }
+    return code;
+  }
+  return getEsImportProxy(id, defaultIsModuleExports);
+}
+
+export function getEsImportProxy(id, defaultIsModuleExports) {
+  const name = getName(id);
+  const exportsName = `${name}Exports`;
+  const requireModule = `require${capitalize(name)}`;
+  let code =
+    `import { getDefaultExportFromCjs } from "${HELPERS_ID}";\n` +
+    `import { __require as ${requireModule} } from ${JSON.stringify(id)};\n` +
+    `var ${exportsName} = ${requireModule}();\n` +
+    `export { ${exportsName} as __moduleExports };`;
+  if (defaultIsModuleExports) {
+    code += `\nexport { ${exportsName} as default };`;
+  } else {
+    code += `export default /*@__PURE__*/getDefaultExportFromCjs(${exportsName});`;
+  }
+  return {
+    code,
+    syntheticNamedExports: '__moduleExports'
+  };
 }
