@@ -4,18 +4,19 @@ import { statSync } from 'fs';
 import { dirname, resolve, sep } from 'path';
 
 import {
-  DYNAMIC_JSON_PREFIX,
-  DYNAMIC_PACKAGES_ID,
-  DYNAMIC_REGISTER_SUFFIX,
+  DYNAMIC_MODULES_ID,
+  ENTRY_SUFFIX,
+  ES_IMPORT_SUFFIX,
   EXPORTS_SUFFIX,
   EXTERNAL_SUFFIX,
   HELPERS_ID,
+  IS_WRAPPED_COMMONJS,
   isWrappedId,
   MODULE_SUFFIX,
   PROXY_SUFFIX,
-  REQUIRE_SUFFIX,
   unwrapId,
-  wrapId
+  wrapId,
+  WRAPPED_SUFFIX
 } from './helpers';
 
 function getCandidatesForExtension(resolved, extension) {
@@ -29,91 +30,99 @@ function getCandidates(resolved, extensions) {
   );
 }
 
-export default function getResolveId(extensions) {
-  function resolveExtensions(importee, importer) {
-    // not our problem
-    if (importee[0] !== '.' || !importer) return undefined;
+export function resolveExtensions(importee, importer, extensions) {
+  // not our problem
+  if (importee[0] !== '.' || !importer) return undefined;
 
-    const resolved = resolve(dirname(importer), importee);
-    const candidates = getCandidates(resolved, extensions);
+  const resolved = resolve(dirname(importer), importee);
+  const candidates = getCandidates(resolved, extensions);
 
-    for (let i = 0; i < candidates.length; i += 1) {
-      try {
-        const stats = statSync(candidates[i]);
-        if (stats.isFile()) return { id: candidates[i] };
-      } catch (err) {
-        /* noop */
-      }
+  for (let i = 0; i < candidates.length; i += 1) {
+    try {
+      const stats = statSync(candidates[i]);
+      if (stats.isFile()) return { id: candidates[i] };
+    } catch (err) {
+      /* noop */
     }
-
-    return undefined;
   }
 
-  return function resolveId(importee, rawImporter, resolveOptions) {
-    if (isWrappedId(importee, MODULE_SUFFIX) || isWrappedId(importee, EXPORTS_SUFFIX)) {
-      return importee;
+  return undefined;
+}
+
+export default function getResolveId(extensions) {
+  return async function resolveId(importee, importer, resolveOptions) {
+    // We assume that all requires are pre-resolved
+    const customOptions = resolveOptions.custom;
+    if (customOptions && customOptions['node-resolve'] && customOptions['node-resolve'].isRequire) {
+      return null;
     }
-
-    const importer =
-      rawImporter && isWrappedId(rawImporter, DYNAMIC_REGISTER_SUFFIX)
-        ? unwrapId(rawImporter, DYNAMIC_REGISTER_SUFFIX)
-        : rawImporter;
-
-    // Except for exports, proxies are only importing resolved ids,
-    // no need to resolve again
-    if (importer && isWrappedId(importer, PROXY_SUFFIX)) {
-      return importee;
-    }
-
-    const isProxyModule = isWrappedId(importee, PROXY_SUFFIX);
-    const isRequiredModule = isWrappedId(importee, REQUIRE_SUFFIX);
-    let isModuleRegistration = false;
-
-    if (isProxyModule) {
-      importee = unwrapId(importee, PROXY_SUFFIX);
-    } else if (isRequiredModule) {
-      importee = unwrapId(importee, REQUIRE_SUFFIX);
-
-      isModuleRegistration = isWrappedId(importee, DYNAMIC_REGISTER_SUFFIX);
-      if (isModuleRegistration) {
-        importee = unwrapId(importee, DYNAMIC_REGISTER_SUFFIX);
-      }
+    if (isWrappedId(importee, WRAPPED_SUFFIX)) {
+      return unwrapId(importee, WRAPPED_SUFFIX);
     }
 
     if (
+      importee.endsWith(ENTRY_SUFFIX) ||
+      isWrappedId(importee, MODULE_SUFFIX) ||
+      isWrappedId(importee, EXPORTS_SUFFIX) ||
+      isWrappedId(importee, PROXY_SUFFIX) ||
+      isWrappedId(importee, ES_IMPORT_SUFFIX) ||
+      isWrappedId(importee, EXTERNAL_SUFFIX) ||
       importee.startsWith(HELPERS_ID) ||
-      importee === DYNAMIC_PACKAGES_ID ||
-      importee.startsWith(DYNAMIC_JSON_PREFIX)
+      importee === DYNAMIC_MODULES_ID
     ) {
       return importee;
+    }
+
+    if (importer) {
+      if (
+        importer === DYNAMIC_MODULES_ID ||
+        // Proxies are only importing resolved ids, no need to resolve again
+        isWrappedId(importer, PROXY_SUFFIX) ||
+        isWrappedId(importer, ES_IMPORT_SUFFIX) ||
+        importer.endsWith(ENTRY_SUFFIX)
+      ) {
+        return importee;
+      }
+      if (isWrappedId(importer, EXTERNAL_SUFFIX)) {
+        // We need to return null for unresolved imports so that the proper warning is shown
+        if (!(await this.resolve(importee, importer, { skipSelf: true }))) {
+          return null;
+        }
+        // For other external imports, we need to make sure they are handled as external
+        return { id: importee, external: true };
+      }
     }
 
     if (importee.startsWith('\0')) {
       return null;
     }
 
-    return this.resolve(
-      importee,
-      importer,
-      Object.assign({}, resolveOptions, {
-        skipSelf: true,
-        custom: Object.assign({}, resolveOptions.custom, {
-          'node-resolve': { isRequire: isProxyModule || isRequiredModule }
-        })
-      })
-    ).then((resolved) => {
-      if (!resolved) {
-        resolved = resolveExtensions(importee, importer);
-      }
-      if (resolved && isProxyModule) {
-        resolved.id = wrapId(resolved.id, resolved.external ? EXTERNAL_SUFFIX : PROXY_SUFFIX);
-        resolved.external = false;
-      } else if (resolved && isModuleRegistration) {
-        resolved.id = wrapId(resolved.id, DYNAMIC_REGISTER_SUFFIX);
-      } else if (!resolved && (isProxyModule || isRequiredModule)) {
-        return { id: wrapId(importee, EXTERNAL_SUFFIX), external: false };
-      }
+    // If this is an entry point or ESM import, we need to figure out if the importee is wrapped and
+    // if that is the case, we need to add a proxy.
+    const resolved =
+      (await this.resolve(importee, importer, Object.assign({ skipSelf: true }, resolveOptions))) ||
+      resolveExtensions(importee, importer, extensions);
+    // Make sure that even if other plugins resolve again, we ignore our own proxies
+    if (
+      !resolved ||
+      resolved.external ||
+      resolved.id.endsWith(ENTRY_SUFFIX) ||
+      isWrappedId(resolved.id, ES_IMPORT_SUFFIX)
+    ) {
       return resolved;
-    });
+    }
+    const moduleInfo = await this.load(resolved);
+    if (resolveOptions.isEntry) {
+      moduleInfo.moduleSideEffects = true;
+      // We must not precede entry proxies with a `\0` as that will mess up relative external resolution
+      return resolved.id + ENTRY_SUFFIX;
+    }
+    const {
+      meta: { commonjs: commonjsMeta }
+    } = moduleInfo;
+    if (commonjsMeta && commonjsMeta.isCommonJS === IS_WRAPPED_COMMONJS) {
+      return { id: wrapId(resolved.id, ES_IMPORT_SUFFIX), meta: { commonjs: { resolved } } };
+    }
+    return resolved;
   };
 }
