@@ -1,20 +1,30 @@
-export default function validateRollupVersion(rollupVersion, peerDependencyVersion) {
-  const [major, minor] = rollupVersion.split('.').map(Number);
-  const versionRegexp = /\^(\d+\.\d+)\.\d+/g;
+export default function validateVersion(actualVersion, peerDependencyVersion, name) {
+  const versionRegexp = /\^(\d+\.\d+\.\d+)/g;
   let minMajor = Infinity;
   let minMinor = Infinity;
+  let minPatch = Infinity;
   let foundVersion;
   // eslint-disable-next-line no-cond-assign
   while ((foundVersion = versionRegexp.exec(peerDependencyVersion))) {
-    const [foundMajor, foundMinor] = foundVersion[1].split('.').map(Number);
+    const [foundMajor, foundMinor, foundPatch] = foundVersion[1].split('.').map(Number);
     if (foundMajor < minMajor) {
       minMajor = foundMajor;
       minMinor = foundMinor;
+      minPatch = foundPatch;
     }
   }
-  if (major < minMajor || (major === minMajor && minor < minMinor)) {
+  if (!actualVersion) {
     throw new Error(
-      `Insufficient Rollup version: "@rollup/plugin-commonjs" requires at least rollup@${minMajor}.${minMinor} but found rollup@${rollupVersion}.`
+      `Insufficient ${name} version: "@rollup/plugin-commonjs" requires at least ${name}@${minMajor}.${minMinor}.${minPatch}.`
+    );
+  }
+  const [major, minor, patch] = actualVersion.split('.').map(Number);
+  if (
+    major < minMajor ||
+    (major === minMajor && (minor < minMinor || (minor === minMinor && patch < minPatch)))
+  ) {
+    throw new Error(
+      `Insufficient ${name} version: "@rollup/plugin-commonjs" requires at least ${name}@${minMajor}.${minMinor}.${minPatch} but found ${name}@${actualVersion}.`
     );
   }
 }
