@@ -1,8 +1,8 @@
 /* eslint-disable import/prefer-default-export */
 
-import { basename, dirname, extname } from 'path';
+import { basename, dirname, extname, relative } from 'path';
 
-import { makeLegalIdentifier } from '@rollup/pluginutils';
+import { createFilter, makeLegalIdentifier } from '@rollup/pluginutils';
 
 export function deconflict(scopes, globals, identifier) {
   let i = 1;
@@ -34,10 +34,32 @@ export function normalizePathSlashes(path) {
   return path.replace(/\\/g, '/');
 }
 
-const VIRTUAL_PATH_BASE = '/$$rollup_base$$';
-export const getVirtualPathForDynamicRequirePath = (path, commonDir) => {
-  const normalizedPath = normalizePathSlashes(path);
-  return normalizedPath.startsWith(commonDir)
-    ? VIRTUAL_PATH_BASE + normalizedPath.slice(commonDir.length)
-    : normalizedPath;
-};
+export const getVirtualPathForDynamicRequirePath = (path, commonDir) =>
+  `/${normalizePathSlashes(relative(commonDir, path))}`;
+
+export function capitalize(name) {
+  return name[0].toUpperCase() + name.slice(1);
+}
+
+export function getStrictRequiresFilter({ strictRequires }) {
+  switch (strictRequires) {
+    case true:
+      return { strictRequiresFilter: () => true, detectCyclesAndConditional: false };
+    // eslint-disable-next-line no-undefined
+    case undefined:
+    case 'auto':
+    case 'debug':
+    case null:
+      return { strictRequiresFilter: () => false, detectCyclesAndConditional: true };
+    case false:
+      return { strictRequiresFilter: () => false, detectCyclesAndConditional: false };
+    default:
+      if (typeof strictRequires === 'string' || Array.isArray(strictRequires)) {
+        return {
+          strictRequiresFilter: createFilter(strictRequires),
+          detectCyclesAndConditional: false
+        };
+      }
+      throw new Error('Unexpected value for "strictRequires" option.');
+  }
+}
