@@ -1,10 +1,8 @@
 import * as commonjsHelpers from "_commonjsHelpers.js";
 import { __exports as input } from "\u0000fixtures/form/constant-template-literal/input.js?commonjs-exports"
-import "\u0000tape?commonjs-require";
-import require$$0 from "\u0000tape?commonjs-proxy";
+import require$$0 from "\u0000CWD/fixtures/form/constant-template-literal/tape.js?commonjs-proxy";
 
 var foo = require$$0;
 console.log(foo);
 
-export default input;
-export { input as __moduleExports };
+export { input as __moduleExports, input as default };
