@@ -4,5 +4,4 @@ import { __exports as input } from "\u0000fixtures/form/defaultIsModuleExports-f
 var _default = input.default = 2;
 var named = input.named = 3;
 
-export default input;
-export { input as __moduleExports, named };
+export { input as __moduleExports, named, input as default };
