@@ -3,5 +3,4 @@ import { __exports as input } from "\u0000fixtures/form/defaultIsModuleExports-f
 
 var named = input.named = 3;
 
-export default input;
-export { input as __moduleExports, named };
+export { input as __moduleExports, named, input as default };
