@@ -5,5 +5,4 @@ input.__esModule = true;
 var _default = input.default = 2;
 var named = input.named = 3;
 
-export default input;
-export { input as __moduleExports, named };
+export { input as __moduleExports, named, input as default };
