@@ -1,10 +1,8 @@
 import * as commonjsHelpers from "_commonjsHelpers.js";
 import { __exports as input } from "\u0000fixtures/form/ignore-ids-function/input.js?commonjs-exports"
-import "\u0000bar?commonjs-require";
-import require$$0 from "\u0000bar?commonjs-proxy";
+import require$$0 from "\u0000CWD/fixtures/form/ignore-ids-function/bar.js?commonjs-proxy";
 
 var foo = require( 'foo' );
 var bar = require$$0;
 
-export default input;
-export { input as __moduleExports };
+export { input as __moduleExports, input as default };
