@@ -1,12 +1,10 @@
 import * as commonjsHelpers from "_commonjsHelpers.js";
 import { __exports as input } from "\u0000fixtures/form/multiple-var-declarations-b/input.js?commonjs-exports"
-import "\u0000./a?commonjs-require";
-import require$$0 from "\u0000./a?commonjs-proxy";
+import require$$0 from "\u0000CWD/fixtures/form/multiple-var-declarations-b/a.js?commonjs-proxy";
 
 var a = require$$0
   , b = 42;
 
 console.log( a, b );
 
-export default input;
-export { input as __moduleExports };
+export { input as __moduleExports, input as default };
