@@ -1,7 +1,6 @@
 import * as commonjsHelpers from "_commonjsHelpers.js";
 import { __exports as input } from "\u0000fixtures/form/multiple-var-declarations-c/input.js?commonjs-exports"
-import "\u0000./b?commonjs-require";
-import require$$0 from "\u0000./b?commonjs-proxy";
+import require$$0 from "\u0000CWD/fixtures/form/multiple-var-declarations-c/b.js?commonjs-proxy";
 
 var a = 'a'
   , b = require$$0
@@ -9,5 +8,4 @@ var a = 'a'
 
 console.log( a, b, c );
 
-export default input;
-export { input as __moduleExports };
+export { input as __moduleExports, input as default };
