@@ -1,7 +1,6 @@
 import * as commonjsHelpers from "_commonjsHelpers.js";
 import { __exports as input_1 } from "\u0000fixtures/form/no-exports-entry/input.js?commonjs-exports"
-import "\u0000./dummy?commonjs-require";
-import require$$0 from "\u0000./dummy?commonjs-proxy";
+import require$$0 from "\u0000CWD/fixtures/form/no-exports-entry/dummy.js?commonjs-proxy";
 
 var dummy = require$$0;
 
@@ -11,5 +10,4 @@ var foo = function () {
 
 var input = 42;
 
-export default input_1;
-export { input_1 as __moduleExports };
+export { input_1 as __moduleExports, input_1 as default };
