@@ -1,7 +1,6 @@
 import * as commonjsHelpers from "_commonjsHelpers.js";
 import { __exports as input } from "\u0000fixtures/form/require-collision/input.js?commonjs-exports"
-import "\u0000foo?commonjs-require";
-import require$$1 from "\u0000foo?commonjs-proxy";
+import require$$1 from "\u0000CWD/fixtures/form/require-collision/foo.js?commonjs-proxy";
 
 (function() {
   var foo = require$$1;
@@ -9,5 +8,4 @@ import require$$1 from "\u0000foo?commonjs-proxy";
   console.log(foo);
 })();
 
-export default input;
-export { input as __moduleExports };
+export { input as __moduleExports, input as default };
