@@ -1,6 +1,4 @@
 import * as commonjsHelpers from "_commonjsHelpers.js";
-import { __exports as input } from "\u0000fixtures/form/unambiguous-with-named-export/input.js?commonjs-exports"
-import "\u0000./foo.js?commonjs-require";
-import "\u0000./foo.js?commonjs-proxy";
+import "\u0000CWD/fixtures/form/unambiguous-with-named-export/foo.js?commonjs-proxy";
 
 export {};
