@@ -18,14 +18,7 @@ module.exports = {
         buildEnd() {
           assert.deepStrictEqual(resolveIdArgs, [
             ['other.js', 'main.js', { custom: { test: 42 }, isEntry: true }],
-            [
-              'other.js',
-              'main.js',
-              // This is the important one
-              { custom: { test: 42, 'node-resolve': { isRequire: false } }, isEntry: true }
-            ],
-            ['main.js', void 0, { custom: {}, isEntry: true }],
-            ['main.js', void 0, { custom: { 'node-resolve': { isRequire: false } }, isEntry: true }]
+            ['main.js', void 0, { custom: {}, isEntry: true }]
           ]);
         },
         resolveId(source, importer, options) {
