@@ -1,4 +1,8 @@
 t.is(
+  require('custom-module')(),
+  '/fixtures/function/dynamic-require-resolve-reference/node_modules/custom-module2'
+);
+t.is(
   require('custom-module2')(),
-  '/$$rollup_base$$/fixtures/function/dynamic-require-resolve-reference/node_modules/custom-module'
+  '/fixtures/function/dynamic-require-resolve-reference/node_modules/custom-module'
 );
