@@ -3,7 +3,7 @@ const property = require('./property');
 const arrayPattern = require('./array-pattern.js');
 const assignmentPattern = require('./assignment-pattern.js');
 
-t.deepEqual(identifier, {});
-t.deepEqual(property, {});
-t.deepEqual(arrayPattern, {});
-t.deepEqual(assignmentPattern, {});
+t.deepEqual(identifier, {}, 'identifier');
+t.deepEqual(property, {}, 'property');
+t.deepEqual(arrayPattern, {}, 'arrayPattern');
+t.deepEqual(assignmentPattern, {}, 'assignmentPattern');
