@@ -1,3 +1,3 @@
 import foo from './foo.js';
 
-t.is(foo, 1);
+t.is(foo, 'require detected');
